Try to fix everything

+++ src/fheroes2/maps/maps_tiles_quantity.cpp
@@ -132,7 +132,7 @@ Spell Maps::Tiles::QuantitySpell( void ) const
 {
     switch ( GetObject( false ) ) {
     case MP2::OBJ_ARTIFACT:
-        return Spell( QuantityVariant() == 15 ? quantity1 : Spell::NONE );
+        return Spell( QuantityVariant() == 15 ? quantity1 : static_cast<int>( Spell::NONE ) );
 
     case MP2::OBJ_SHRINE1:
     case MP2::OBJ_SHRINE2:
